Remove codes related with BlockPool
