--- blockchain/infra/adapter/event_handler_test.go
@@ -1,90 +1,0 @@
-/*
- * Copyright 2018 It-chain
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package adapter_test
-
-import (
-	"testing"
-	"time"
-
-	"github.com/it-chain/engine/blockchain"
-	"github.com/it-chain/engine/blockchain/infra/adapter"
-	"github.com/it-chain/engine/blockchain/test/mock"
-	"github.com/magiconair/properties/assert"
-)
-
-func TestEventHandler_HandleBlockAddToPoolEvent(t *testing.T) {
-	tests := map[string]struct {
-		input struct {
-			blockchain.BlockAddToPoolEvent
-		}
-		err error
-	}{
-		"success": {
-			input: struct {
-				blockchain.BlockAddToPoolEvent
-			}{BlockAddToPoolEvent: blockchain.BlockAddToPoolEvent{
-				Seal:      []byte{0x1},
-				PrevSeal:  []byte{0x1},
-				Height:    uint64(12),
-				TxList:    []byte{0x1},
-				TxSeal:    [][]byte{{0x1}},
-				Timestamp: time.Now(),
-				Creator:   []byte{0x1},
-			}},
-			err: nil,
-		},
-		"event w/o height test": {
-			input: struct {
-				blockchain.BlockAddToPoolEvent
-			}{BlockAddToPoolEvent: blockchain.BlockAddToPoolEvent{
-				Seal:      []byte{0x1},
-				PrevSeal:  []byte{0x1},
-				TxList:    []byte{0x1},
-				TxSeal:    [][]byte{{0x1}},
-				Timestamp: time.Now(),
-				Creator:   []byte{0x1},
-			}},
-			err: adapter.ErrBlockMissingProperties,
-		},
-		"event w/o seal, prevseal test": {
-			input: struct {
-				blockchain.BlockAddToPoolEvent
-			}{BlockAddToPoolEvent: blockchain.BlockAddToPoolEvent{
-				TxList:    []byte{0x1},
-				TxSeal:    [][]byte{{0x1}},
-				Timestamp: time.Now(),
-				Creator:   []byte{0x1},
-			}},
-			err: adapter.ErrBlockMissingProperties,
-		},
-	}
-
-	blockApi := mock.BlockApi{}
-	blockApi.CheckAndSaveBlockFromPoolFunc = func(height blockchain.BlockHeight) error {
-		assert.Equal(t, height, uint64(12))
-		return nil
-	}
-	eventHandler := adapter.NewEventHandler(blockApi)
-
-	for testName, test := range tests {
-		t.Logf("running test case %s", testName)
-
-		err := eventHandler.HandleBlockAddToPoolEvent(test.input.BlockAddToPoolEvent)
-
-		assert.Equal(t, err, test.err)
-	}
-}
